+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Added `DropDownItemSearchable` interface to allow the `T` of `DropDownItem<T>` to determine the way items are searched.
 * Added `toString` method to `DropDownItem` to use for `DropDownItem.build` and `DropDownItem.satisfiesSearch`.
 * Updated language example to showcase new customization abilities.
+* Renamed some internal variables and utilized new `select`/`deselect`/`selectAll`/`deselectAll` methods.
 
 ## 0.0.4